Move firebase init to avoid crash when resuming from Android background
The root cause is that an Android app sometimes will not kill itself after
people click hareware back buttn. So the result is JS class App will be
instantialized twice in the same lifecycle. If we put firebase.initializeApp
in the constructor and it will throw exception about duplicate firebase apps.